Improve CmsisDapProtocol class
Remove CmsisDapCore class and move the enums CommandId and IdInfo to
CmsisDapProtocol
